+++ src/projection/drawing.hpp
@@ -4,8 +4,8 @@
 
 namespace projection {
   class drawing {
-    int _w = -1;
-    int _h = -1;
+    float _w = -1;
+    float _h = -1;
     float _elapsed = 0.0f;
     bool _initialized = false;
     uint64_t _frames = 0;
@@ -13,14 +13,14 @@ namespace projection {
   public:
     explicit drawing();
 
-    int width() { return _w; };
-    int height() { return _h; };
+    float width() { return _w; };
+    float height() { return _h; };
     float elapsed() { return _elapsed; };
     bool initialized() { return _initialized; };
     uint64_t frames() { return _frames; };
     float frame_rate() {  return _frames / _elapsed; };
 
-    void initialize(int width, int height);
+    void initialize(float width, float height);
     virtual void setup() =0;
 
     void redraw();
